LayoutTests/media/video-loopcount.html

<video src=content/test.mp4 playcount=3 loopend=0.2s end=0.2s controls></video>
<p>Test that playcount attribute causes playback to end after right number of repeats.</p>
<script src=video-test.js></script>
<script>
    testExpected("video.playCount", 3);
    testExpected("video.currentLoop", 0);

    waitForEvent('ended', function () { 
        testExpected("video.currentLoop", 2);
        endTest();
    } );

    video.play();
</script>
