LayoutTests/media/video-loopcount-expected.txt

Test that playcount attribute causes playback to end after right number of repeats.

EXPECTED (video.playCount == '3') OK
EXPECTED (video.currentLoop == '0') OK
EVENT(ended)
EXPECTED (video.currentLoop == '2') OK
END OF TEST
